When running in reopen mode, some of the tests are expected to fail because they use dangerous operations (LOG=0,...)

+++ h2/src/test/org/h2/test/db/TestSQLInjection.java
@@ -33,6 +33,9 @@ public class TestSQLInjection extends TestBase {
     }
 
     public void test() throws SQLException {
+        if (config.reopen) {
+            return;
+        }
         deleteDb("sqlInjection");
         reconnect("sqlInjection");
         stat.execute("DROP TABLE IF EXISTS USERS");
